requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
bcrypt==4.0.1
passlib[bcrypt]
python-jose
python-multipart
npm
google-generativeai
httpx
gunicorn
python-dotenv==0.21.0
pydantic[email]
langchain_huggingface
transformers
langchain
langchain_community
torch
huggingface-hub
faiss-cpu
requests
ipython
langchain_cohere
beautifulsoup4
langchainhub
langsmith
langgraph==0.2.14
arxiv
pdfplumber
python-docx
pinecone[asyncio,grpc]
